+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/index.tsx
@@ -17,8 +17,8 @@ import {
   ItemTitle,
   VersionContainer,
 } from './elements';
-import ConfigurationFiles from './items/ConfigurationFilesOld';
-import { ConfigurationFiles as ConfigurationFilesNew } from './items/ConfigurationFiles';
+import ConfigurationFiles from './items/ConfigurationFiles';
+import { ConfigurationFiles as ConfigurationFilesNew } from './screens/ConfigurationFiles';
 import { Deployment } from './items/Deployment';
 import { FilesItem } from './items/Files';
 import { GitHub } from './items/GitHub';
