+++ packages/app/src/app/pages/Sandbox/Editor/Workspace/items/ConfigurationFiles/index.tsx
@@ -1,55 +1,96 @@
 import React from 'react';
-import {
-  Element,
-  Collapsible,
-  Text,
-  Button,
-  Stack,
-  Grid as BaseGrid,
-} from '@codesandbox/components/lib/';
 import getDefinition from '@codesandbox/common/lib/templates';
+import getUI from '@codesandbox/common/lib/templates/configuration/ui';
+import { Module, Configuration } from '@codesandbox/common/lib/types';
 import { resolveModule } from '@codesandbox/common/lib/sandbox/modules';
 
 import { useOvermind } from 'app/overmind';
-import styled, { withTheme } from 'styled-components';
-import { TypescriptIcon } from 'app/components/TypescriptIcon';
+
+import BookIcon from 'react-icons/lib/md/library-books';
+import UIIcon from 'react-icons/lib/md/dvr';
+
+import Tooltip from '@codesandbox/common/lib/components/Tooltip';
+
+import { Description, WorkspaceSubtitle } from '../../elements';
 import {
-  NetlifyIcon,
-  PrettierIcon,
-  NPMIcon,
-  ZeitIcon,
-  JSIcon,
-  CodeSandboxIcon,
-} from './Icons';
+  FilesContainer,
+  File,
+  FileTitle,
+  FileDescription,
+  CreateButton,
+} from './elements';
 
-const getIcon = name => {
-  const icons = {
-    'netlify.toml': NetlifyIcon,
-    'package.json': NPMIcon,
-    'now.json': ZeitIcon,
-    '.prettierrc': PrettierIcon,
-    'jsconfig.json': JSIcon,
-    'tsconfig.json': TypescriptIcon,
-    'sandbox.config.json': CodeSandboxIcon,
+type FileConfigProps = {
+  path: string;
+  info: {
+    module?: Module;
+    config: Configuration;
   };
-
-  return icons[name] || JSIcon;
+  createModule?: (title: string) => void;
+  openModule?: (id: string) => void;
 };
 
-const Grid = styled(BaseGrid)`
-  grid-template-columns: 1fr 100px;
-  grid-gap: ${({ theme }) => theme.space[4]}px;
-`;
+const FileConfig = ({
+  info,
+  path,
+  createModule,
+  openModule,
+}: FileConfigProps) => {
+  const { module, config } = info;
+  return (
+    <File
+      created={Boolean(module)}
+      key={path}
+      onClick={
+        openModule
+          ? () => {
+              openModule(module.id);
+            }
+          : undefined
+      }
+    >
+      <FileTitle>
+        {config.title}{' '}
+        <Tooltip content="More Info">
+          <a
+            href={config.moreInfoUrl}
+            target="_blank"
+            rel="noreferrer noopener"
+            title="Documentation"
+            style={{ marginLeft: '.25rem' }}
+          >
+            <BookIcon />
+          </a>
+        </Tooltip>
+        {getUI(config.type) && (
+          <Tooltip content="Editable with UI">
+            <UIIcon style={{ marginLeft: '.5rem' }} />
+          </Tooltip>
+        )}
+      </FileTitle>
+      <FileDescription>{config.description}</FileDescription>
+      {!module && (
+        <CreateButton
+          onClick={() => {
+            // TODO make this support nested paths (create dir etc)
+            createModule(info.config.title);
+          }}
+        >
+          Create File
+        </CreateButton>
+      )}
+    </File>
+  );
+};
 
-export const ConfigurationFilesComponent = ({ theme }) => {
+const ConfigurationFiles = () => {
   const {
-    state: {
-      editor: { currentSandbox },
-    },
+    state,
     actions: { files, editor },
   } = useOvermind();
 
-  const { configurationFiles } = getDefinition(currentSandbox.template);
+  const sandbox = state.editor.currentSandbox;
+  const { configurationFiles } = getDefinition(sandbox.template);
 
   const createdPaths = {};
   const restPaths = {};
@@ -60,11 +101,7 @@ export const ConfigurationFilesComponent = ({ theme }) => {
       const config = configurationFiles[p];
 
       try {
-        const module = resolveModule(
-          p,
-          currentSandbox.modules,
-          currentSandbox.directories
-        );
+        const module = resolveModule(p, sandbox.modules, sandbox.directories);
         createdPaths[p] = {
           config,
           module,
@@ -77,89 +114,48 @@ export const ConfigurationFilesComponent = ({ theme }) => {
     });
 
   return (
-    <>
-      <Collapsible title="Configuration Files" defaultOpen>
-        <Stack
-          direction="vertical"
-          gap={6}
-          style={{ padding: `0 ${theme.space[3]}px` }}
-        >
-          <Element>
-            <Text as="div" marginBottom={2}>
-              Configuration your Sandbox
-            </Text>
-            <Text variant="muted">
-              Codesandbox supports several config giles per template, you van
-              see and edit all supported files for the current sandbox here.
-            </Text>
-          </Element>
-          <Stack direction="vertical" gap={4}>
-            {Object.keys(createdPaths).map(path => {
-              const { config, module } = createdPaths[path];
-              const Icon = getIcon(config.title);
-              return (
-                <Element>
-                  <Stack gap={2} marginBottom={2}>
-                    <Icon />
-                    <Text size={2}>{config.title}</Text>
-                  </Stack>
-                  <Grid>
-                    <Text size={2} variant="muted">
-                      {config.description}
-                    </Text>
-                    <Button
-                      style={{ width: 'auto' }}
-                      variant="secondary"
-                      onClick={() => editor.moduleSelected({ id: module.id })}
-                    >
-                      Edit
-                    </Button>
-                  </Grid>
-                </Element>
-              );
-            })}
-          </Stack>
-        </Stack>
-      </Collapsible>
-      <Collapsible title="Other Configuration" defaultOpen>
-        <Stack
-          direction="vertical"
-          gap={4}
-          style={{ padding: `0 ${theme.space[3]}px` }}
-        >
-          {Object.keys(restPaths).map(path => {
-            const { config } = restPaths[path];
-            const Icon = getIcon(config.title);
-            return (
-              <Element>
-                <Stack gap={2} marginBottom={2}>
-                  <Icon />
-                  <Text size={2}>{config.title}</Text>
-                </Stack>
-                <Grid>
-                  <Text size={2} variant="muted">
-                    {config.description}
-                  </Text>
-                  <Button
-                    style={{ width: 'auto' }}
-                    variant="secondary"
-                    onClick={() =>
-                      files.moduleCreated({
-                        title: config.title,
-                        directoryShortid: null,
-                      })
-                    }
-                  >
-                    Create File
-                  </Button>
-                </Grid>
-              </Element>
-            );
-          })}
-        </Stack>
-      </Collapsible>
-    </>
+    <div>
+      <Description>
+        CodeSandbox supports several config files per template, you can see and
+        edit all supported files for the current sandbox here.
+      </Description>
+
+      <FilesContainer>
+        <WorkspaceSubtitle>Existing Configurations</WorkspaceSubtitle>
+        {Object.keys(createdPaths).map(path => {
+          const info = createdPaths[path];
+
+          return (
+            <FileConfig
+              key={path}
+              openModule={id => {
+                editor.moduleSelected({ id });
+              }}
+              path={path}
+              info={info}
+            />
+          );
+        })}
+        {!!Object.keys(restPaths).length && (
+          <WorkspaceSubtitle>Other Configurations</WorkspaceSubtitle>
+        )}
+        {Object.keys(restPaths).map(path => {
+          const info = restPaths[path];
+
+          return (
+            <FileConfig
+              key={path}
+              createModule={title => {
+                files.moduleCreated({ title, directoryShortid: null });
+              }}
+              path={path}
+              info={info}
+            />
+          );
+        })}
+      </FilesContainer>
+    </div>
   );
 };
 
-export const ConfigurationFiles = withTheme(ConfigurationFilesComponent);
+export default ConfigurationFiles;
